Removing the server instance
Upon thinking on this further, I don't want to have the Websocket server being able to control anything involving Supabase. The front end is the one that should be manipulating the data and communicating with Supabase. The Websocket server is only really listening for what the user and the key listener emit.